requirements.txt
scipy
numpy
matplotlib
pandas~=1.3.4
scikit-learn~=1.0.1
umap-learn
umap-learn[plot]
